// deno-http-server/enum.ts
export const enum TESTNET_MNEMONIC {
    ROCOCO =
      "where sign course conduct include wide because skull boss slice close tomorrow",
    SHIBUYA =
      "acoustic satisfy keen evidence hat citizen version magic nuclear misery invest tag",
  }
  
  export const enum RPC_ENDPOINT {
    ROCOCO = "wss://rococo-contracts-rpc.polkadot.io:443",
    SHIBUYA = "wss://shibuya-rpc.dwellir.com:443",
  }
  
  export const enum SS58_FORMAT {
    POLKADOT = 0,
    KUSAMA = 2,
    SHIBUYA = 5,
    DEFAULT = 42,
  }
  
  export const enum DECIMAL_PLACE_FROM_PLANCK {
    POLKADOT = 10,
    ROCOCO = 12,
    SHIBUYA = 18,
  }